Move HeterogeneousCore/AlpakaCore/interface/MakerMacros.h under alpaka/
